+++ src/components/UsersAdmin/__tests__/UsersAdmin-test.js
@@ -8,7 +8,7 @@ import RoleStore from '../../../stores/RoleStore';
 import UserStore from '../../../stores/UserStore';
 import RoleActions from '../../../actions/RoleActions';
 import UserActions from '../../../actions/UserActions';
-import UsersAdmin from '../index.jsx';
+import UsersAdmin from '../UsersAdmin.jsx';
 
 describe('UsersAdmin', function() {
